% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm_configure.R
\name{tm_configure}
\alias{tm_configure}
\title{Store Twilio SID, Twilio token, Twilio phone number, and target phone number for later use.}
\usage{
tm_configure(twilio_sid = NULL, twilio_token = NULL,
  twilio_phone_number = NULL, target_phone_number = NULL,
  overwrite = FALSE)
}
\arguments{
\item{twilio_sid}{Your Twilio SID.}

\item{twilio_token}{Your Twilio Token.}

\item{twilio_phone_number}{Your Twilio phone number.}

\item{target_phone_number}{The phone number you want to receive a text message at}

\item{overwrite}{By default, this is set to \code{FALSE}. If \code{TRUE}, R will overwrite existing .Renviron variables that match the ones named in this function.}
}
\value{
A .Renviron file that stores the supplied paramaters for ease of use in future R sessions.
}
\description{
Store Twilio SID, Twilio token, Twilio phone number, and target phone number for later use.
}
\examples{

# set parameters
tm_configure(twilio_sid   = "AC1ffb497954c02119a0f0720784bde131",
             twilio_token = "f5c8af81ddadaf3e8904b909f975c369",
             twilio_phone_number = "1234567890",
             target_phone_number = "0987654321")

# view and edit the .Renvion file
file.edit("~/.Renviron")

}
